Add fake mode to purge_orphaned_uploads. Tests. Closes #620.

+++ panda/management/commands/purge_orphaned_uploads.py
@@ -6,12 +6,23 @@ import os
 from django.conf import settings
 from django.core.management.base import BaseCommand
 
+from optparse import make_option
 from panda.models import DataUpload, RelatedUpload
 
 class Command(BaseCommand):
-    help = 'Audit uploads and local files; delete any not associated with a dataset'
+    help = 'Audit uploads and local files, deleting any not associated with a dataset.'
+    option_list = BaseCommand.option_list + (
+        make_option('--fake',
+            action='store_true',
+            dest='fake',
+            default=False,
+            help='Only describe what files would be deleted, don\'t actually delete them.'),
+        )
 
     def handle(self, *args, **options):
+        if options['fake']:
+            self.stdout.write('Running in fake mode! No files will actually be deleted!')
+
         local_files = os.listdir(settings.MEDIA_ROOT)
         data_uploads = DataUpload.objects.all()
         related_uploads = RelatedUpload.objects.all()
@@ -24,11 +35,18 @@ class Command(BaseCommand):
                 pass
 
             if not upload.dataset:
-                self.stdout.write('Deleted upload: %s\n' % upload)
-                upload.delete()
+                if options['fake']:
+                    self.stdout.write('Would delete upload: %s\n' % upload)
+                else:
+                    self.stdout.write('Deleted upload: %s\n' % upload)
+                    upload.delete()
 
         for f in local_files:
             path = os.path.join(settings.MEDIA_ROOT, f)
-            self.stdout.write('Deleted file: %s\n' % path)
-            os.remove(path)
+
+            if options['fake']:
+                self.stdout.write('Would delete file: %s\n' % path)
+            else:
+                self.stdout.write('Deleted file: %s\n' % path)
+                os.remove(path)
